requirements.txt
python-dotenv==1.0.1
rank-bm25==0.2.2
langchain-groq==0.2.0
langchain-community==0.3.3
jq==1.8.0
